+++ web/ce/components/workspace-notifications/index.ts
@@ -1,1 +1,1 @@
-export * from "./root";
+export * from "./notification-card/root";
